Asteroids own rotation
The individual rotation of the steroids on their own axes has been completed.
The rotation script consists of two checkboxes available on the interface, where it is allowed to configure, without the need to open the script, the individual rotation of each prefab, as well as its direction.
The natural rotation of the asteroids was also programmed to be generated in a minimum random space close to the number of speed informed, making the rotation of several asteroids on the screen more random/caotic and natural.