Asteroid orbits
The asteroids' orbits were duly programmed with gizmos to facilitate their modification.

BezierPath functionality:
Create a new empty object and apply this script
>Route
Now create four empty child objects inside the main object
>ControlPoint1
>ControlPoint2
>ControlPoint3
>ControlPoint4
The main object will be used as a route to the asteroids
The four children will be used as a guide for the route drawn the bezier curves
Set the size to four on the interface. Put the child in the right order inside the interface.
Mould the bezier curve on the game scene