Asteroid_movement finished
This script is responsible for move the asteroid throw the Bezier curve path.
In order for the asteroid to travel over more than one Bézier turn on its route. Add two "route" gaps to the game scene, connect the endpoint of the first with the second and so on. Then, specify in the interface the number of routes that the asteroid will follow and assimilate the necessary "route" objects in the correct  order.
To make the asteroid move cyclically, make sure that the last point on the route is at the same origin as the first.